pyproject.toml
[tool.poetry]
name = "join-group"
version = "0.1.0"
description = ""
authors = ["jeff3071 <jeff563715@gmail.com>"]
readme = "README.md"
packages = [{include = "join_group"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2.6"
djangorestframework = "^3.14.0"
django-ninja = "^0.22.2"
djangorestframework-simplejwt = "^5.3.0"
psycopg2-binary = "^2.9.9"
django-cors-headers = "^4.3.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
